Consolidate sitemap controllers
