+++ Tests/AppTests/SitemapTests.swift
@@ -149,10 +149,10 @@ class SitemapTests: SnapshotTestCase {
             .query(on: app.db, owner: "owner", repository: "repo0")
 
         // MUT
-        let sitemap = try await SiteMapView.package(owner: packageResult.repository.owner,
-                                                    repository: packageResult.repository.name,
-                                                    lastActivityAt: packageResult.repository.lastActivityAt,
-                                                    linkablePathUrls: [])
+        let sitemap = try await SiteMapController.package(owner: packageResult.repository.owner,
+                                                          repository: packageResult.repository.name,
+                                                          lastActivityAt: packageResult.repository.lastActivityAt,
+                                                          linkablePathUrls: [])
         let xml = sitemap.render(indentedBy: .spaces(2))
 
         // Validation
@@ -178,10 +178,10 @@ class SitemapTests: SnapshotTestCase {
         ]
 
         // MUT
-        let sitemap = try await SiteMapView.package(owner: packageResult.repository.owner,
-                                                    repository: packageResult.repository.name,
-                                                    lastActivityAt: packageResult.repository.lastActivityAt,
-                                                    linkablePathUrls: linkablePathUrls)
+        let sitemap = try await SiteMapController.package(owner: packageResult.repository.owner,
+                                                          repository: packageResult.repository.name,
+                                                          lastActivityAt: packageResult.repository.lastActivityAt,
+                                                          linkablePathUrls: linkablePathUrls)
         let xml = sitemap.render(indentedBy: .spaces(2))
 
         // Validation
